refactor: rename details movie ui state
